// webcomponents/studio/src/types/slot-type.ts
export enum SlotType {
  SECTION = 'section',
  H1 = 'h1',
  H2 = 'h2',
  H3 = 'h3',
  OL = 'ol',
  UL = 'ul',
  HR = 'deckgo-hr',
  IMG = 'deckgo-lazy-img',
  CODE = 'deckgo-highlight-code',
  SOCIAL = 'deckgo-social',
  REVEAL = 'deckgo-reveal',
  REVEAL_LIST = 'deckgo-reveal-list',
  DRAG_RESIZE_ROTATE = 'deckgo-drr',
  DEMO = 'deckgo-demo',
  MATH = 'deckgo-math',
  WORD_CLOUD = 'deckgo-word-cloud',
  MARKDOWN = 'deckgo-markdown'
}